Eliminate liferay-theme-deps-7.x packages
In this step we remove all the packages except for
"liferay-theme-deps-normalize" which I will do in the next commit; for
now I am just calling it from the "install" script of the generated
theme package.

Note that I am not explicitly adding a dependency on gulp 3.9.1 here
because it is a transitive dependency of gulp-sass anyway, and in any
case, we have to update to gulp v4 shortly anyway. At that point, I may
add it back just to be explicit.

Related: https://github.com/liferay/liferay-themes-sdk/issues/168