Remove generator dependence on globally installed gulp instance (#172)
If you try to generate a theme without a globally installed copy of gulp
you'll get `Error: spawn gulp ENOENT` at the end. We do tell people (in
the README) to install gulp globally, but that shouldn't be necessary:
it is already a dependency of the generator and we can just use the
generator's copy.

Test plan: `yo ./packages/generator-liferay-theme` without a globally
installed copy of `gulp` and see it work without blowing up.

Closes: https://github.com/liferay/liferay-themes-sdk/issues/172