Remove scary warnings from normalize process in the generator
Before this commit, when you created a new 7.1 (or 7.2) theme, you would
see this warning in the console:

    Unable to add inject tags to /Users/greghurrell/code/liferay-themes-sdk/zzz-theme/node_modules/liferay-frontend-theme-unstyled/templates/portal_normal.vm

This sounds like an error, but it's totally expected because the theme
in 7.1 doesn't have any Velocity templates.

So, this commit reworks things so that you just see non-scary status
messages instead.

In 7.1+ you see:

    Completed tag injection for node_modules/liferay-frontend-theme-unstyled/templates/portal_normal.ftl
    Skipping tag injection for node_modules/liferay-frontend-theme-unstyled/templates/portal_normal.vm (does not exist)

In 7.0 you see:

    Completed tag injection for node_modules/liferay-frontend-theme-unstyled/templates/portal_normal.ftl
    Completed tag injection for node_modules/liferay-frontend-theme-unstyled/templates/portal_normal.vm