WIP Try to get build tests working again
All out custom magic with the `require.resolve` and friends is gone now,
or almost all of it, but these tests are failing still, in part because
Jest is a weird environment with a fake `require` implementation. So I'm
finding that it doesn't behavior well outside of the project root (like
in the temporary directories where we are testing themes out); various
hacks in this commit to try and work around that, including an `npm
install` to get the deps locally which is never going to cut it because
it adds about 10 or 15 seconds to the test run. Parking this in the PR
for now anyway and will resume tomorrow.